+++ packages/experimental/Checkbox/src/Checkbox.types.ts
@@ -24,11 +24,12 @@ export interface CheckboxTokens extends FontTokens, IForegroundColorTokens, IBac
    * States that can be applied to a checkbox
    */
   disabled?: CheckboxTokens;
-  boxAtEnd?: CheckboxTokens;
+  labelIsBefore?: CheckboxTokens;
   hovered?: CheckboxTokens;
   focused?: CheckboxTokens;
   pressed?: CheckboxTokens;
   checked?: CheckboxTokens;
+  circular?: CheckboxTokens;
 }
 
 export interface CheckboxProps extends Omit<IViewProps, 'onPress'> {
@@ -39,6 +40,13 @@ export interface CheckboxProps extends Omit<IViewProps, 'onPress'> {
   checked?: boolean;
 
   /**
+   * Allows you to set the checkbox to have circular styling.
+   *
+   * @platform Android, iOS, windows, win32
+   */
+  circular?: boolean;
+
+  /**
    * Default checked state. Mutually exclusive to ‘checked’. Use this if you want an uncontrolled component, and
    * want the Checkbox instance to maintain its own state.
    */
@@ -46,8 +54,10 @@ export interface CheckboxProps extends Omit<IViewProps, 'onPress'> {
 
   /**
    * Allows you to set the checkbox to be at the before (start) or after (end) the label
+   *
+   * @default after
    */
-  boxSide?: 'start' | 'end';
+  labelPosition?: 'before' | 'after';
 
   /**
    * Disabled state of the checkbox.
@@ -87,9 +97,9 @@ export interface CheckboxState extends IPressableState {
   disabled?: boolean;
 
   /**
-   * Determines position of Checkbox. True if boxSide='end'
+   * Determines position of Checkbox. True if labelPosition is set to 'before'
    */
-  boxAtEnd?: boolean;
+  labelIsBefore?: boolean;
 }
 
 export interface CheckboxInfo {