+++ packages/experimental/Button/CHANGELOG.md
@@ -1,8 +1,61 @@
 # Change Log - @fluentui-react-native/experimental-button
 
-This log was last generated on Tue, 21 Dec 2021 20:56:31 GMT and should not be manually modified.
+This log was last generated on Fri, 07 Jan 2022 21:27:06 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.11.14
+
+Fri, 07 Jan 2022 21:27:06 GMT
+
+### Patches
+
+- Fix styling order (ruaraki@microsoft.com)
+
+## 0.11.13
+
+Thu, 06 Jan 2022 23:14:38 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/experimental-activity-indicator to v0.5.8
+- Bump @fluentui-react-native/experimental-text to v0.7.9
+- Bump @fluentui-react-native/framework to v0.7.9
+- Bump @fluentui-react-native/icon to v0.11.3
+- Bump @fluentui-react-native/interactive-hooks to v0.11.11
+- Bump @fluentui-react-native/theme-tokens to v0.11.4
+- Bump @fluentui-react-native/tokens to v0.11.4
+
+## 0.11.12
+
+Thu, 06 Jan 2022 21:30:08 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/experimental-activity-indicator to v0.5.7
+- Bump @fluentui-react-native/experimental-text to v0.7.8
+- Bump @fluentui-react-native/framework to v0.7.8
+- Bump @fluentui-react-native/icon to v0.11.2
+- Bump @fluentui-react-native/interactive-hooks to v0.11.10
+- Bump @fluentui-react-native/theme-tokens to v0.11.3
+
+## 0.11.11
+
+Wed, 05 Jan 2022 23:53:54 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/interactive-hooks to v0.11.9
+
+## 0.11.10
+
+Tue, 04 Jan 2022 23:05:12 GMT
+
+### Patches
+
+- Add checks for mutually exclusive props (ruaraki@microsoft.com)
+- Fix spacing around icon (ruaraki@microsoft.com)
+- Bump @fluentui-react-native/styling-utils to v0.3.2
 
 ## 0.11.9