+++ packages/experimental/MenuButton/CHANGELOG.md
@@ -1,8 +1,56 @@
 # Change Log - @fluentui-react-native/experimental-menu-button
 
-This log was last generated on Thu, 23 Dec 2021 02:03:45 GMT and should not be manually modified.
+This log was last generated on Fri, 07 Jan 2022 21:27:06 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.2.18
+
+Fri, 07 Jan 2022 21:27:06 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/experimental-button to v0.11.14
+
+## 0.2.17
+
+Thu, 06 Jan 2022 23:14:38 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/contextual-menu to v0.14.9
+- Bump @fluentui-react-native/experimental-button to v0.11.13
+- Bump @fluentui-react-native/framework to v0.7.9
+- Bump @fluentui-react-native/tokens to v0.11.4
+
+## 0.2.16
+
+Thu, 06 Jan 2022 21:30:08 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/contextual-menu to v0.14.8
+- Bump @fluentui-react-native/experimental-button to v0.11.12
+- Bump @fluentui-react-native/framework to v0.7.8
+
+## 0.2.15
+
+Wed, 05 Jan 2022 23:53:54 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/contextual-menu to v0.14.7
+- Bump @fluentui-react-native/experimental-button to v0.11.11
+
+## 0.2.14
+
+Tue, 04 Jan 2022 23:05:12 GMT
+
+### Patches
+
+- Fix tests (ruaraki@microsoft.com)
+- Fix tests (ruaraki@microsoft.com)
+- Bump @fluentui-react-native/experimental-button to v0.11.10
 
 ## 0.2.13