+++ packages/libraries/core/CHANGELOG.md
@@ -1,8 +1,73 @@
 # Change Log - @fluentui/react-native
 
-This log was last generated on Thu, 23 Dec 2021 02:03:45 GMT and should not be manually modified.
+This log was last generated on Thu, 06 Jan 2022 23:14:38 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.27.15
+
+Thu, 06 Jan 2022 23:14:38 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/button to v0.21.11
+- Bump @fluentui-react-native/callout to v0.19.14
+- Bump @fluentui-react-native/checkbox to v0.14.11
+- Bump @fluentui-react-native/contextual-menu to v0.14.9
+- Bump @fluentui-react-native/focus-trap-zone to v0.8.11
+- Bump @fluentui-react-native/focus-zone to v0.8.11
+- Bump @fluentui-react-native/interactive-hooks to v0.11.11
+- Bump @fluentui-react-native/link to v0.15.11
+- Bump @fluentui-react-native/menu-button to v0.7.15
+- Bump @fluentui-react-native/persona to v0.12.9
+- Bump @fluentui-react-native/persona-coin to v0.10.9
+- Bump @fluentui-react-native/pressable to v0.8.11
+- Bump @fluentui-react-native/radio-group to v0.14.11
+- Bump @fluentui-react-native/separator to v0.11.9
+- Bump @fluentui-react-native/tabs to v0.6.11
+- Bump @fluentui-react-native/text to v0.11.6
+
+## 0.27.14
+
+Thu, 06 Jan 2022 21:30:08 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/button to v0.21.10
+- Bump @fluentui-react-native/callout to v0.19.13
+- Bump @fluentui-react-native/checkbox to v0.14.10
+- Bump @fluentui-react-native/contextual-menu to v0.14.8
+- Bump @fluentui-react-native/focus-trap-zone to v0.8.10
+- Bump @fluentui-react-native/focus-zone to v0.8.10
+- Bump @fluentui-react-native/interactive-hooks to v0.11.10
+- Bump @fluentui-react-native/link to v0.15.10
+- Bump @fluentui-react-native/menu-button to v0.7.14
+- Bump @fluentui-react-native/persona to v0.12.8
+- Bump @fluentui-react-native/persona-coin to v0.10.8
+- Bump @fluentui-react-native/pressable to v0.8.10
+- Bump @fluentui-react-native/radio-group to v0.14.10
+- Bump @fluentui-react-native/separator to v0.11.8
+- Bump @fluentui-react-native/tabs to v0.6.10
+- Bump @fluentui-react-native/text to v0.11.5
+
+## 0.27.13
+
+Wed, 05 Jan 2022 23:53:54 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/button to v0.21.9
+- Bump @fluentui-react-native/callout to v0.19.12
+- Bump @fluentui-react-native/checkbox to v0.14.9
+- Bump @fluentui-react-native/contextual-menu to v0.14.7
+- Bump @fluentui-react-native/focus-trap-zone to v0.8.9
+- Bump @fluentui-react-native/focus-zone to v0.8.9
+- Bump @fluentui-react-native/interactive-hooks to v0.11.9
+- Bump @fluentui-react-native/link to v0.15.9
+- Bump @fluentui-react-native/menu-button to v0.7.13
+- Bump @fluentui-react-native/pressable to v0.8.9
+- Bump @fluentui-react-native/radio-group to v0.14.9
+- Bump @fluentui-react-native/tabs to v0.6.9
 
 ## 0.27.12