+++ packages/utils/interactive-hooks/CHANGELOG.md
@@ -1,8 +1,32 @@
 # Change Log - @fluentui-react-native/interactive-hooks
 
-This log was last generated on Tue, 21 Dec 2021 20:56:31 GMT and should not be manually modified.
+This log was last generated on Thu, 06 Jan 2022 23:14:38 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.11.11
+
+Thu, 06 Jan 2022 23:14:38 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/icon to v0.11.3
+
+## 0.11.10
+
+Thu, 06 Jan 2022 21:30:08 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/icon to v0.11.2
+
+## 0.11.9
+
+Wed, 05 Jan 2022 23:53:54 GMT
+
+### Patches
+
+- Change export method (ruaraki@microsoft.com)
 
 ## 0.11.8